font_get_size + emSize fix from UTMT

+++ src/data_win.h
@@ -399,7 +399,7 @@ typedef struct {
     bool present;
     const char* name;
     const char* displayName;
-    uint32_t emSize;
+    float emSize;
     bool bold;
     bool italic;
     uint16_t rangeStart;
